Move event handlers from UI to RepositoryMonitor
