+++ RepoZ.Api/RepoZ.Api.csproj
@@ -38,9 +38,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\IErrorHandler.cs" />
+    <Compile Include="Common\IThreadDispatcher.cs" />
     <Compile Include="Common\SingularityQueue.cs" />
     <Compile Include="Git\DefaultRepositoryInformationAggregator.cs" />
-    <Compile Include="Git\IRepositoryCache.cs" />
     <Compile Include="Git\IRepositoryStore.cs" />
     <Compile Include="Git\IRepositoryInformationAggregator.cs" />
     <Compile Include="Git\IRepositoryObserverFactory.cs" />
