+++ RepoZ.UI.Win.Wpf/RepoZ.UI.Win.Wpf.csproj
@@ -77,6 +77,7 @@
     <Reference Include="PresentationFramework" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="WpfThreadDispatcher.cs" />
     <Page Include="App.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
